+++ src/history/js/history-hash.js
@@ -324,56 +324,48 @@ Y.extend(HistoryHash, HistoryBase, {
 // YUIDoc.
 
 /**
- * <p>
- * Synthetic <code>window.onhashchange</code> event that normalizes differences
- * across browsers and provides support for browsers that don't natively support
- * <code>onhashchange</code>.
- * </p>
- *
- * <p>
- * This event is provided by the <code>history-hash</code> module.
- * </p>
- *
- * <p>
- * <strong>Usage example:</strong>
- * </p>
- *
- * <code><pre>
- * YUI().use('history-hash', function (Y) {
- * &nbsp;&nbsp;Y.on('hashchange', function (e) {
- * &nbsp;&nbsp;&nbsp;&nbsp;// Handle hashchange events on the current window.
- * &nbsp;&nbsp;}, Y.config.win);
- * });
- * </pre></code>
- *
- * @event hashchange
- * @param {EventFacade} e Event facade with the following additional
- *   properties:
- *
- * <dl>
- *   <dt>oldHash</dt>
- *   <dd>
- *     Previous hash fragment value before the change.
- *   </dd>
- *
- *   <dt>oldUrl</dt>
- *   <dd>
- *     Previous URL (including the hash fragment) before the change.
- *   </dd>
- *
- *   <dt>newHash</dt>
- *   <dd>
- *     New hash fragment value after the change.
- *   </dd>
- *
- *   <dt>newUrl</dt>
- *   <dd>
- *     New URL (including the hash fragment) after the change.
- *   </dd>
- * </dl>
- * @for YUI
- * @since 3.2.0
- */
+Synthetic <code>window.onhashchange</code> event that normalizes differences
+across browsers and provides support for browsers that don't natively support
+<code>onhashchange</code>.
+
+This event is provided by the <code>history-hash</code> module.
+
+@example
+
+    YUI().use('history-hash', function (Y) {
+      Y.on('hashchange', function (e) {
+        // Handle hashchange events on the current window.
+      }, Y.config.win);
+    });
+
+@event hashchange
+@param {EventFacade} e Event facade with the following additional
+  properties:
+
+<dl>
+  <dt>oldHash</dt>
+  <dd>
+    Previous hash fragment value before the change.
+  </dd>
+
+  <dt>oldUrl</dt>
+  <dd>
+    Previous URL (including the hash fragment) before the change.
+  </dd>
+
+  <dt>newHash</dt>
+  <dd>
+    New hash fragment value after the change.
+  </dd>
+
+  <dt>newUrl</dt>
+  <dd>
+    New URL (including the hash fragment) after the change.
+  </dd>
+</dl>
+@for YUI
+@since 3.2.0
+**/
 
 hashNotifiers = GlobalEnv._notifiers;